Migrate insertion sort
- Move implementation to header file
- Add missing curly braces
- Update README
- Delete old insertion_sort.cpp
